// src/core/types/pokemon-types.ts
export type PokemonTypes =
  "bug"
  | "dark"
  | "dragon"
  | "electric"
  | "fairy"
  | "fighting"
  | "fire"
  | "flying"
  | "ghost"
  | "grass"
  | "ground"
  | "ice"
  | "normal"
  | "poison"
  | "psychic"
  | "rock"
  | "steel"
  | "water"
  | "none";

export interface PokemonConfigInterface {
  background: string;
  color: string;
}